+++ pulse_lib/tests/channel_timing/test_delays.py
@@ -35,6 +35,35 @@ def test():
 
     return None
 
+def test2():
+    pulse = context.init_pulselib(n_gates=2, n_sensors=2,
+                                  rf_sources=True)
+
+    s = pulse.mk_segment()
+
+    s.P1.wait(100)
+    s.P1.add_block(0, 20, 100)
+    s.P2.add_block(0, 20, -100)
+    s.SD1.acquire(0, 100)
+    s.SD2.acquire(0, 100)
+
+    # now with other delays
+    pulse.add_channel_delay('P1', 0)
+    pulse.add_channel_delay('SD1', 0)
+    pulse.add_channel_delay('SD2', 0)
+    pulse.digitizer_channels['SD2'].rf_source.mode = 'continuous'
+    pulse.digitizer_channels['SD2'].rf_source.startup_time_ns = 0
+    pulse.digitizer_channels['SD2'].rf_source.prolongation_ns = 0
+
+    sequence = pulse.mk_sequence([s])
+    sequence.n_rep = 10
+    context.plot_awgs(sequence, print_acquisitions=True)
+
+    return None
+
+#%%
+
 
 if __name__ == '__main__':
     ds = test()
+    ds2 = test2()